HARP-5820: Add globe support to WebTileDataSource.
Also, add examples demonstrating how to render satellites and web tiles
with sphere projection.
